add ouq_misc, ytrue-directed surrogate examples (#265)

+++ examples3/xrd_optML4xs.py
@@ -16,15 +16,19 @@ from ouq_models import WrapModel, InterpModel
 from emulators import cost4 as cost, x4 as target, bounds4 as bounds
 from mystic.cache.archive import file_archive, read as get_db
 
-# remove any prior cached evaluations of truth (i.e. an 'expensive' model)
-if os.path.exists("truth"):
-    import shutil
-    shutil.rmtree("truth")
+# prepare truth (i.e. an 'expensive' model)
+nx = 4; ny = None
+#archive = get_db('truth.db', type=file_archive)
+truth = WrapModel('truth', cost, nx=nx, ny=ny, cached=False)#archive)
+
+# remove any prior cached evaluations of truth
+import shutil
+if os.path.exists("truth"): shutil.rmtree("truth")
+if os.path.exists("error.txt"): os.remove("error.txt")
+if os.path.exists("log.txt"): os.remove("log.txt")
 
 # remove any prior cached evaluations of surrogate
-if os.path.exists("surrogate"):
-    import shutil
-    shutil.rmtree("surrogate")
+if os.path.exists("surrogate"): shutil.rmtree("surrogate")
 
 try: # parallel maps
     from pathos.maps import Map
@@ -34,8 +38,6 @@ except ImportError:
     pmap = None
 
 # generate a training dataset by sampling truth
-#archive = get_db('truth.db', type=file_archive)
-truth = WrapModel('truth', cost, nx=4, ny=None, cached=False)#archive)
 data = truth.sample(bounds, pts=[2, 1, 1, 1], pmap=map)#, archive=archive)
 
 # shutdown mapper
@@ -43,7 +45,7 @@ if pmap is not None:
     pmap.close(); pmap.join(); pmap.clear()
 
 # create an inexpensive surrogate for truth
-surrogate = InterpModel("surrogate", nx=4, ny=None, data=truth, smooth=0.0,
+surrogate = InterpModel("surrogate", nx=nx, ny=ny, data=truth, smooth=0.0,
                         noise=0.0, method="thin_plate", extrap=False)
 
 # iterate until error (of candidate minimum) < 1e-3
@@ -52,7 +54,7 @@ import numpy as np
 tracker = LoggingMonitor(1, filename='error.txt', label='error')
 from mystic.abstract_solver import AbstractSolver
 from mystic.termination import VTR
-loop = AbstractSolver(4) # nx
+loop = AbstractSolver(nx)
 loop.SetTermination(VTR(1e-3)) #XXX: VTRCOG, TimeLimits, etc?
 loop.SetEvaluationLimits(maxiter=500)
 loop.SetGenerationMonitor(tracker)
@@ -67,6 +69,8 @@ while not loop.Terminated():
 
     # evaluate truth at the same input as the surrogate critical points
     ytrue = list(map(truth, surr.coords))
+    # add most recent candidate extrema to truth database
+    data.load(surr.coords, ytrue)
 
     # compute absolute error between truth and surrogate at candidate extrema
     idx = np.argmin(ytrue)
@@ -78,12 +82,10 @@ while not loop.Terminated():
     print("evaluations of truth: %s" % len(data))
 
     # save to tracker if less than current best
-    if len(tracker) and tracker.y[-1] < error[idx]:
+    ysave = error # track error when learning surrogate
+    if len(tracker) and tracker.y[-1] < ysave[idx]:
         tracker(*tracker[-1])
-    else: tracker(surr.coords[idx], error[idx])
-
-    # add most recent candidate extrema to truth database
-    data.load(surr.coords, ytrue)
+    else: tracker(surr.coords[idx], ysave[idx])
 
 # get the results at the best parameters from the truth database
 xbest = tracker[-1][0]
